login: unify pw_shell script test
At first, a blank in pw_shell is sufficient to assume that the user
login shell is actually a shell script.

If execution fails, the test consists of testing argv[0] for "/bin/sh"
which is set in case of shell script.

But "/bin/sh" can also be the actual login shell set for the user.
If execution fails, the error message erroneously talks about a shell
script.

Keep a boolean variable around to perform the test exactly once to
have a precise error message.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>