lsfd: (refactor) make the steps for new_file consistent
In parse_maps_line of the original code, we called readlink after stat
for gathering arguments that were passed to new_file().

The order was different from that in collect_file_symlink.
In parse_maps_line, we called stat after readlink.

The difference of the order may make the way of handling
errors of readlink and stat complicated.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>