treewide: fix newlines when using fputs
The `fputs()` introduced in #2561 does not automatically include a
trailing newline which the original `puts()` did.
This broke the `--help` output of various utilities.

Changing all format strings to include a newline is not feasible as it
would force all translations to be adapted.

Instead introduce a new helper `fputsln()` that works as expected.