loopdev: Fix race between systemd-udevd and delete loop device losetup -d
We have found a race between systemd-udevd and "losetup -d" command
when we try to delete a loop device using "losetup -d" while other tools
like blkid/mkfs are run. The systemd-udevd may get invoked (after
blkid/mkfs are done) to process the rules as it watches all the block
device nodes.

When "losetup -d" is started while systemd-udevd is in progress, the
kernel code of "losetup -d", loop_clr_fd() will get EBUSY while rescanning
partitions as the block device is already opened by systemd-udevd.
Due to this EBUSY error, the partitions are not rescanned but the loop
device is deleted causing in stale partitions of loop device remaining.

So, while deleting the loop device, the operations must be synchronized.
If the systemd-udev is started first, the other commands should wait.
If the "losetup -d" started first, then systemd-udev must be stopped from
processing the rules so that device and its partitions are deleted
properly. Fortunately the systemd-udev takes the flock on the loop device
in shared mode LOCK_SH but the "losetup -d" command doesn't take any lock.
So, to avoid this race, "losetup -d" should take flock in LOCK_EX mode.

Signed-off-by: Gulam Mohamed <gulam.mohamed@oracle.com>