meson: Remove lingering mq_libs variable
The cherry-pick of #2879 in PR #2941 to the stable/v2.40 branch didn't
include the removal of mq_libs like it should have.
I must have resolved a conflict incorrectly in the cherry-pick.
This just removes the lingering definitions which are no longer used.
The extra find_library call for rt is a problem.

Signed-off-by: Jordan Williams <jordan@jwillikers.com>