cal: fix --week use and colors
* don't overwrite --week=<number> by the current date, default only
  to the current year, if the year is not specified

* introduce "weeks" color sequence to colorize all week numbers.

Note that the man page suggests using "weeknumber" to colorize all
week numbers, but the code only uses this sequence for the week
requested by the --week option. This commit maintains backward
compatibility, although it would be better to use "weeknumber" for all
week numbers and "weekwanted" for the desired week.

Signed-off-by: Karel Zak <kzak@redhat.com>