meson: use a / b instead of join_paths(a, b)
According to [1], since meson 0.49.0, using the / operator on strings
is equivalent to calling join_paths().

As the expressions written using the / operator are shorter and nicer
than their equivalents written using join_paths(), and given that
we've been using this new operator since commit
5f2c862a4aff19f5bc732e677405f04e2d07c211, let's convert remaining
join_paths() invocations to use the / operator instead.

[1] https://mesonbuild.com/Reference-manual_functions.html#join_paths