env: save parsed variables into ul_env_list
The struct ul_env_list is currently only used to store the value of
env when sanitizing environ[]. It would be beneficial to also use the
functions related to ul_env_list for other purposes. In order to do
this, it is important to keep the NAME and VALUE components separate.

* Accept non-terminated name and value for env_list_add().

* Add env_list_add_from_string() to add item to the list from a
  "name=value" string.

* Add env_list_add_variable() to add item to the list using separate
  "name" and "value" strings.

* Update __sanitize_env() to use env_list_add_from_string().

* Modify the test to use "abc" as the value instead of copying the
  variable name (with '=') to the value.

Signed-off-by: Karel Zak <kzak@redhat.com>