nsenter: add functions to enable/disable namespaces
Currently, enabled namespaces are those with an open file descriptor.
However, if we support pidfd, this will become unnecessary and we will
need an FD-independent enable/disable mechanism.

It also makes sense to delay opening --target <pid> namespaces files
until everything is ready and only handle it in one place.

Signed-off-by: Karel Zak <kzak@redhat.com>