test_mkfds: reserve file descriptors in the early stage of execution
A factory specified with command line opens some files.  After
opening, the factory remaps the opened file descriptors (ofds) to file
descriptors (rfds) specified with the command line with dup2 system all.

This remapping may fail if there is an overlap between ofds and rfds.
With this change, there cannot be an overlap between ofds and rfds;
test_mkfds reserves rfds in the early stage of execution.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>