libmount: improving readability
The function name mnt_optlist_is_recursive() is not accurate. It is
only used for propagation operations based on the MS_REC flag and
should not be confused with any other recursive stuff, such as
ro=recursive.

A more appropriate name for this function would be mnt_optlist_is_rpropagation().

Signed-off-by: Karel Zak <kzak@redhat.com>