findmnt: improve reliability of match testing
The "match" is defined if a string or void data is defined for the
matching function. It is necessary to test both variants instead of
relying solely on get_match(), which only returns string data.

This patch introduces the is_defined_match() macro to hide this detail
in the code and make it more readable.

Signed-off-by: Karel Zak <kzak@redhat.com>